+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * fix exception when outputLimit is used [#776](https://github.com/jcubic/jquery.terminal/issues/776)
 * fix showing links for emails like prompt [#768](https://github.com/jcubic/jquery.terminal/issues/768)
 * fix line wrapping when text between formatting have brackets [#772](https://github.com/jcubic/jquery.terminal/issues/772)
+* fix flickr of dynamic prompt [#757](https://github.com/jcubic/jquery.terminal/issues/757)
 
 ## 2.32.1
 ### Bugfix